Update Full App
AS of right now the shiny app can only process tsv files that originate from the same paper. Functionality must be added to the app so that different paper dataframes can be stored and processed separately before ultimately being joined in a PCA plot post COMBAT.
In a non shiny environment this would be easy enough, just run the functions multiple times and save the results in different variables, but in shiny all these created variables should be accessible to the user.
The initial idea is to ask the user the disease or title of the paper before any computation is done. All subsequent matrices must then be saved separately. In each subsequent page the app should then ask the user again what disease should be processed during that step and pull up the appropriate data. This can be achieved by having reactive inputSelects() although there is a still a little work to be done to keep these different origin dataframes independent of one another.
This approach seems like it would require the user to repeat a lot of actions to get the whole way through. more thought should be put into this problem to come up with an elegant solution. Ask feedback from team